+++ changelog.md
@@ -487,8 +487,7 @@ are 9001 for HSQLDB, 9092 for H2 and 1527 for Derby.
 - Reduce serialization when saving assignment paths
 - Use compact arrays for the graph and heap
 - Faster to vehicles conversion
-- Improved map navigation responsiveness
- 
+- Zoom-dependent simplification for the Mercator projection